updated main.py - added a short sleep after nmap
after the nmap scan sometimes it would turn the display off and then immediately back on, I guess it was checking for "latency" before the whole nmap output had even appeared. So I added a short sleep after the scan, before the find statements which fixed this up for me.